requirements.txt
awscli
black
flake8
ipython
isort
jupyterlab
matplotlib
mkdocs
notebook
numpy
pandas
pip
python-dotenv
scikit-learn
Django==5.1.5
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
xgboost
grad-cam
transformers
shap
lime
datasets
accelerate>=0.26.0
psycopg2-binary
drf-yasg